Extract ActionListener interface
