+++ src/main/java/org/sagebionetworks/web/client/widget/entity/menu/v2/ActionListener.java
@@ -1,0 +1,10 @@
+package org.sagebionetworks.web.client.widget.entity.menu.v2;
+
+public interface ActionListener {
+	/**
+	 * Called when the users selects an action.
+	 *
+	 * @param action The selected action.
+	 */
+	void onAction(Action action);
+}
