+++ src/main/java/org/sagebionetworks/web/client/widget/entity/menu/v2/ActionView.java
@@ -1,7 +1,7 @@
 package org.sagebionetworks.web.client.widget.entity.menu.v2;
 
 import org.gwtbootstrap3.client.ui.base.HasIcon;
-import org.sagebionetworks.web.client.widget.entity.menu.v2.ActionMenuWidget.ActionListener;
+
 import com.google.gwt.user.client.ui.HasEnabled;
 import com.google.gwt.user.client.ui.HasText;
 import com.google.gwt.user.client.ui.HasVisibility;
@@ -32,7 +32,7 @@ public interface ActionView extends IsWidget, HasText, HasIcon, HasEnabled, HasV
 	/**
 	 * Add an action listener to this view.
 	 * 
-	 * @param listner
+	 * @param listener
 	 */
 	void addActionListener(ActionListener listener);
