+++ src/test/java/org/sagebionetworks/web/unitclient/widget/subscription/SubscribeButtonWidgetTest.java
@@ -33,6 +33,7 @@ import org.sagebionetworks.web.client.security.AuthenticationController;
 import org.sagebionetworks.web.client.utils.Callback;
 import org.sagebionetworks.web.client.widget.entity.controller.SynapseAlert;
 import org.sagebionetworks.web.client.widget.entity.menu.v2.Action;
+import org.sagebionetworks.web.client.widget.entity.menu.v2.ActionListener;
 import org.sagebionetworks.web.client.widget.entity.menu.v2.ActionMenuWidget;
 import org.sagebionetworks.web.client.widget.subscription.SubscribeButtonWidget;
 import org.sagebionetworks.web.client.widget.subscription.SubscribeButtonWidgetView;
@@ -68,7 +69,7 @@ public class SubscribeButtonWidgetTest {
 	@Mock
 	ActionMenuWidget mockActionMenuWidget;
 	@Captor
-	ArgumentCaptor<ActionMenuWidget.ActionListener> actionListenerCaptor;
+	ArgumentCaptor<ActionListener> actionListenerCaptor;
 	private static final String TEST_OBJECT_ID = "3";
 	private static final String TEST_SUBSCRIPTION_ID = "8837";
 
@@ -152,7 +153,7 @@ public class SubscribeButtonWidgetTest {
 		verify(mockView).clear();
 		verify(mockSynAlert).clear();
 		verify(mockView).showUnfollowButton();
-		verify(mockActionMenuWidget).setActionListener(eq(Action.FOLLOW), any(ActionMenuWidget.ActionListener.class));
+		verify(mockActionMenuWidget).setActionListener(eq(Action.FOLLOW), any(ActionListener.class));
 		verify(mockActionMenuWidget).setActionText(Action.FOLLOW, "Unfollow Thread");
 	}
 
@@ -163,7 +164,7 @@ public class SubscribeButtonWidgetTest {
 		verify(mockView).clear();
 		verify(mockSynAlert).clear();
 		verify(mockView).showFollowButton();
-		verify(mockActionMenuWidget).setActionListener(eq(Action.FOLLOW), any(ActionMenuWidget.ActionListener.class));
+		verify(mockActionMenuWidget).setActionListener(eq(Action.FOLLOW), any(ActionListener.class));
 		verify(mockActionMenuWidget).setActionText(Action.FOLLOW, "Follow Forum");
 	}
